+++ src/components/Checkbox/variants.js
@@ -1,3 +1,8 @@
 module.exports = {
-  all: color => [`ripple-${color}-500`, `ripple-${color}`, `text-${color}-500`]
+  all: color => [
+    `text-${color}-500`,
+    `hover:bg-${color}-trans`,
+    `bg-${color}-trans`,
+    `hover:bg-${color}-transDark`
+  ]
 };