+++ src/components/List/ListItem.svelte
@@ -1,6 +1,7 @@
 <script>
   import { createEventDispatcher } from "svelte";
   import Icon from "../Icon";
+  import createRipple from "../Ripple/ripple.js";
 
   export let icon = "";
   export let id = "";
@@ -13,8 +14,11 @@
   export let to = "";
   export let selected = false;
   export let tabindex = null;
-  export let basicClasses = "transition p-4 cursor-pointer text-gray-700 flex items-center z-10";
+  export let basicClasses = "relative overflow-hidden my-1 transition p-4 cursor-pointer text-gray-700 flex items-center z-10";
+  export let itemClasses = "";
+  export let selectedClasses = "bg-gray-200";
 
+  const ripple = createRipple();
   const dispatch = createEventDispatcher();
 
   function change() {
@@ -25,10 +29,6 @@
 </script>
 
 <style>
-  li.selected {
-    @apply bg-gray-200;
-  }
-
   li:focus {
     @apply bg-gray-50;
   }
@@ -43,12 +43,10 @@
 </style>
 
 <li
-  class={basicClasses}
+  use:ripple
+  class="{basicClasses} {selected ? selectedClasses : ''}"
   class:navigation
-  class:selected
-  class:hover:bg-gray-300={!navigation}
-  class:ripple-white={navigation}
-  class:ripple-gray={!navigation}
+  class:hover:bg-gray-transLight={!navigation}
   class:py-2={dense}
   class:text-gray-600={disabled}
   {tabindex}
@@ -65,7 +63,7 @@
   {/if}
 
   <div class="flex flex-col p-0">
-    <div>
+    <div class={itemClasses}>
       <slot>{text}</slot>
     </div>
     {#if subheading}