+++ src/components/Tabs/TabButton.svelte
@@ -1,28 +1,39 @@
 <script>
   import Icon from "../Icon";
+  import createRipple from "../Ripple/ripple.js";
+  import utils, { ClassBuilder } from "../../utils/classes.js";
 
   export let icon = "";
   export let id = "";
   export let text = "";
   export let to = "";
   export let selected = "";
-  export let color = "primary-500";
+  export let color = "primary";
 
-  $: isSelected = selected === id;
+  let className = "";
+  export {className as class};
 
-  const classes =
-    "text-center w-40 p-4 cursor-pointer flex mx-auto items-center opacity-75 text-sm h-full";
+  const ripple = createRipple(color);
+
+  const defaultClasses =
+    "relative overflow-hidden text-center w-40 p-4 cursor-pointer flex mx-auto items-center opacity-75 text-sm h-full";
+
+
+  const { txt } = utils(color);
+
+  $: textColor = selected === id ? txt() : "text-white";
 </script>
 
 {#if to}
   <a
+    use:ripple
     href={to}
-    class="{classes} ripple-{color} text-white {isSelected ? `text-${color}` : ''}"
+    class="{className} {defaultClasses} {textColor}"
     on:click
     class:uppercase={icon}>
     <div class="flex flex-col items-center content-center mx-auto">
       {#if icon}
-        <Icon class="mb-1" color={isSelected ? `text-${color}` : ''}>{icon}</Icon>
+        <Icon class="mb-1" color={textColor}>{icon}</Icon>
       {/if}
 
       <div>
@@ -32,15 +43,14 @@
   </a>
 {:else}
   <li
-    class="{classes} ripple-{color} text-white {isSelected ? `text-${color}` : ''}"
-    on:click={() => {
-      selected = id;
-    }}
+    use:ripple
+    class="{className} {defaultClasses} {textColor}"
+    on:click={() => selected = id }
     on:click
     class:uppercase={icon}>
     <div class="flex flex-col items-center content-center mx-auto">
       {#if icon}
-        <Icon class="mb-1" color={isSelected ? `text-${color}` : ''}>{icon}</Icon>
+        <Icon class="mb-1" color={textColor}>{icon}</Icon>
       {/if}
 
       <div>