+++ src/index.js
@@ -21,8 +21,10 @@ import TextField from "./components/TextField";
 import DataTable from "./components/DataTable";
 import Switch from "./components/Switch";
 import Tooltip from "./components/Tooltip";
+import Treeview from "./components/Treeview";
+import Ripple from "./components/Ripple";
 
-export { Scrim, Ripple, Spacer } from "./components/Util";
+export { Scrim, Spacer } from "./components/Util";
 
 export {
   Button,
@@ -42,12 +44,14 @@ export {
   Select,
   Slider,
   Snackbar,
-  Tab, 
+  Tab,
   Tabs,
-  TabButton, 
+  TabButton,
   TextField,
   AppBar,
   DataTable,
+  Treeview,
   Switch,
-  Tooltip
+  Tooltip,
+  Ripple
 };