+++ src/routes/components/tabs.svelte
@@ -1,6 +1,6 @@
 <script>
   import { Tabs, Tab } from "smelte";
-  import Code from "components/Code";
+  import Code from "docs/Code.svelte";
   import { Image } from "smelte";
 
   import tabs from "examples/tabs.txt";
@@ -31,7 +31,11 @@
     color="secondary"
     let:selected
     {loading}
-    items={[{ id: '1', text: 'Cats', icon: 'alarm_on' }, { id: '2', text: 'Kittens', icon: 'bug_report' }, { id: '3', text: 'Kitties', icon: 'eject' }]}>
+    items={[
+      { id: '1', text: 'Cats', icon: 'alarm_on' },
+      { id: '2', text: 'Kittens', icon: 'bug_report' },
+      { id: '3', text: 'Kitties', icon: 'eject' }
+    ]}>
     <div
       slot="content"
       class="flex items-center content-center overflow-hidden w-full bg-gray-900